+++ libs/expo/shared/ui-components/src/index.ts
@@ -1,1 +1,3 @@
 export * from './lib';
+
+export { default as BaseModalLayout } from './lib/BaseModalLayout/BaseModalLayout';
