+++ libs/expo/shared/ui-components/src/lib/BaseModalLayout/index.tsx
@@ -1,0 +1,1 @@
+export { default as BaseModalLayout} from './BaseModalLayout';
